ZED: REMOVED, kicked out, FU Zed.
It started to get on my nerfes.

1.) The config-docs are garbage. They say "look at the Autocomplete to
see what is possible". Great job guys. If your Values are just Fucking
arrays, that only take strings, I still don't know what functions i can
trigger.

2.) the last update, yesterday, deleted an entire folder without prompt
and no way to undo it.
I'm sure it's just a bug but I'm tired of this shit.
Each day, there is a new ZED update, that changes and/or breaks stuff.
How am I supposed to use an Editor, that I can't be sure, will still
work as expected tomorrow.

That folder contained a lot of files, god damnit. Why couldn't you move
it to the trashcan? Nooooo, you had to delete it from the drive directly
instead.